pyproject.toml
[project]
name = "product_feature_heygen_video"
version = "0.1.0"
description = "product_feature_heygen_video using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.130.0,<1.0.0",
    "google-cloud-storage>=2.0.0",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.0.2"
]

[project.scripts]
product_feature_heygen_video = "product_feature_heygen_video.main:run"
run_crew = "product_feature_heygen_video.main:run"
train = "product_feature_heygen_video.main:train"
replay = "product_feature_heygen_video.main:replay"
test = "product_feature_heygen_video.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
